GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - test/controllers/welcome_controller_test.rb
   - app/controllers/welcome_controller.rb
   - test/controllers/dashboard_controller_test.rb
   - app/views/welcome/show.html.erb
   - app/controllers/application_controller.rb
   - app/controllers/dashboard_controller.rb
   - config/routes.rb

Your previous branch was: refs/heads/main

The sha for that commit was: eac901ed3d22dff705ee1694583060cc0fe70503

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch